changed folder structure / moved types to seperate files
