+++ src/components/LocationTriviaScreen/index.tsx
@@ -1,0 +1,2 @@
+import { TEMPDATA } from "../../TEMPDATA";
+export { default } from "./LocationTriviaScreen";
