+++ src/components/TriviaListDisplay/TriviaListDisplay.tsx
@@ -1,15 +1,22 @@
 import React, { useState, useEffect } from "react";
 
+import {ScreenDisplay} from './ScreenDisplay'
 import { TriviaFilterSelector } from "./TriviaFilterSelector";
-import Header from "../layout/Header";
-import {TriviaCard} from './TriviaCard'
-//Should I have another container for the trivia cards?
+import { TriviaCardList } from "./TriviaCardList";
+
+
+
+
+
+
+
+import { TriviaEntry } from "../../types/index";
+import { TriviaFilter } from "../../types/index";
+
+
+
 
 import { TEMPDATA } from "../../TEMPDATA";
-import { BooleanLiteral } from "typescript";
-import { TriviaCardList } from "./TriviaCardList";
-import { LocationCard } from "./LocationCard";
-
 //Seperate Types Folder?
 
 //TriviaFilter
@@ -20,65 +27,6 @@ import { LocationCard } from "./LocationCard";
 //How best to represent time?
 //setup ENUM incase of unknown? for now we do boolean or string
 
-interface TriviaEntry {
-  locationId: number;
-  //For instances where location has more than one trivia night
-  triviaId: number;
-  locationName: string;
-  locationWebsite: string;
-  locationInstagram: string;
-  locationTwitter: string;
-  locationFacebook: string;
-  locationImage: string;
-  locationType: string;
-  locationPhoneNumber: string;
-  locationAddressStreet: string;
-  locationAddressCity: string;
-  locationAddressState: string;
-  locationAddressZip: string;
-  locationAddressFull: string;
-  triviaDay: string;
-  triviaStartTime: number;
-  //On day of trivia
-  locationCloseTime: number;
-  triviaCompany: string | null;
-  locationFood: {
-    foodTruck: boolean | string;
-    kitchen: boolean | string;
-    bringYourOwn: boolean | string;
-  };
-  locationAlcohol: {
-    wine: boolean | string;
-    liquor: boolean | string;
-    beer: boolean | string;
-  };
-  locationAllowsDogs: boolean | string;
-  locationOutdoorSeating: {
-    forTrivia: boolean | string;
-    general: boolean | string;
-  };
-}
-
-
-interface TriviaFilter {
-  time: {
-    sunday: boolean;
-    monday: boolean;
-    tuesday: boolean;
-    wednesday: boolean;
-    thursday: boolean;
-    friday: boolean;
-    saturday: boolean;
-    timeStartBeginRange: number;
-    timeStartEndRange: number;
-  };
-  location: {
-    //Valid Zip Code Detection?
-    zipCode: number | null;
-    mileage: number | null;
-  };
-  searchTerm: string;
-}
 
 
 
@@ -106,7 +54,7 @@ const defaultFilter = {
 //TODO: FILTERS
 //TODO: CLEAN UP CSS
 
-export const TriviaListDisplayContainer: React.FC = () => {
+const TriviaListDisplay: React.FC = () => {
     //Should we have a list of filtered Trivias?
 //   const [filteredTriviaList, setFilteredTriviaList] = useState<Array<TriviaEntry> | null>(null);
 
@@ -165,9 +113,9 @@ export const TriviaListDisplayContainer: React.FC = () => {
             />
           </div>
           {/* RESTRAUNT PAGE -- hidden unless lg*/}
-          
+
           <div className="hidden lg:flex flex-col justify-center h-screen sticky sideFrameHeight self-end bottom-0 z-10 basis-5/12 bg-orange-700">
-            <LocationCard
+            <ScreenDisplay
               triviaEntry={triviaEntryList?.find(
                 (triviaEntry) =>
                   triviaEntry.locationId === currentlySelectedLocation
@@ -179,3 +127,5 @@ export const TriviaListDisplayContainer: React.FC = () => {
     </div>
   );
 };
+
+export default TriviaListDisplay;
