+++ src/components/TriviaListDisplay/TriviaListDisplay.types.tsx
@@ -1,0 +1,1 @@
+export {}
