+++ src/components/TriviaListDisplay/index.tsx
@@ -1,0 +1,1 @@
+export {default} from './TriviaListDisplay'
