+++ src/pages/index.tsx
@@ -1,6 +1,7 @@
 import Head from "next/head";
 import Image from "next/image";
-import {TriviaListDisplayContainer} from "@/components/TriviaListDisplayContainer/TriviaListDisplayContainer";
+
+import TriviaListDisplay from "@/components/TriviaListDisplay";
 
 import { Inter } from "next/font/google";
 
@@ -9,7 +10,7 @@ const inter = Inter({ subsets: ["latin"] });
 export default function Home() {
   return (
     <>
-     <TriviaListDisplayContainer/>
+     <TriviaListDisplay/>
     </>
   );
 }
