+++ src/types/index.ts
@@ -1,0 +1,61 @@
+//any shared types
+export interface TriviaEntry {
+  locationId: number;
+  //For instances where location has more than one trivia night
+  triviaId: number;
+  locationName: string;
+  locationWebsite: string;
+  locationInstagram: string;
+  locationTwitter: string;
+  locationFacebook: string;
+  locationImage: string;
+  locationType: string;
+  locationPhoneNumber: string;
+  locationAddressStreet: string;
+  locationAddressCity: string;
+  locationAddressState: string;
+  locationAddressZip: string;
+  locationAddressFull: string;
+  triviaDay: string;
+  triviaStartTime: number;
+  //On day of trivia
+  locationCloseTime: number;
+  triviaCompany: string | null;
+  locationFood: {
+    foodTruck: boolean | string;
+    kitchen: boolean | string;
+    bringYourOwn: boolean | string;
+  };
+  locationAlcohol: {
+    wine: boolean | string;
+    liquor: boolean | string;
+    beer: boolean | string;
+  };
+  locationAllowsDogs: boolean | string;
+  locationOutdoorSeating: {
+    forTrivia: boolean | string;
+    general: boolean | string;
+  };
+}
+
+
+
+export interface TriviaFilter {
+  time: {
+    sunday: boolean;
+    monday: boolean;
+    tuesday: boolean;
+    wednesday: boolean;
+    thursday: boolean;
+    friday: boolean;
+    saturday: boolean;
+    timeStartBeginRange: number;
+    timeStartEndRange: number;
+  };
+  location: {
+    //Valid Zip Code Detection?
+    zipCode: number | null;
+    mileage: number | null;
+  };
+  searchTerm: string;
+}
